+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- The version command in the CLI now once again correctly identifies if the binary
+  is the regular mycelium or the mycelium-private binary. 
+
 ## [0.7.4] - 2026-03-23
 
 ### Added